feat(Card): internal card with private props
